+++ CoreArrangementPlugins/src/au/gov/asd/tac/constellation/plugins/arrangements/spectral/GraphSpectrumEmbedder.java
@@ -16,16 +16,15 @@
 package au.gov.asd.tac.constellation.plugins.arrangements.spectral;
 
 import au.gov.asd.tac.constellation.graph.GraphReadMethods;
-import java.util.Collection;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.Map;
-import java.util.Set;
 import org.apache.commons.math3.linear.EigenDecomposition;
 import org.apache.commons.math3.linear.MatrixUtils;
+import org.eclipse.collections.api.IntIterable;
 import org.eclipse.collections.api.map.primitive.IntIntMap;
 import org.eclipse.collections.api.map.primitive.MutableIntIntMap;
+import org.eclipse.collections.api.map.primitive.MutableIntObjectMap;
 import org.eclipse.collections.impl.map.mutable.primitive.IntIntHashMap;
+import org.eclipse.collections.impl.map.mutable.primitive.IntObjectHashMap;
+import org.eclipse.collections.impl.set.mutable.primitive.IntHashSet;
 
 /**
  *
@@ -33,15 +32,15 @@ import org.eclipse.collections.impl.map.mutable.primitive.IntIntHashMap;
  */
 public class GraphSpectrumEmbedder {
 
-    public static Map<Integer, double[]> spectralEmbedding(final GraphReadMethods rg, final Set<Integer> includedVertices) {
-        final Map<Integer, double[]> vertexPositions = new HashMap<>();
+    public static MutableIntObjectMap<double[]> spectralEmbedding(final GraphReadMethods rg, final IntIterable includedVertices) {
+        final MutableIntObjectMap<double[]> vertexPositions = new IntObjectHashMap<>();
 
         // Don't position anything if there are fewer than 3 vertices to embedd - this embedding shouldn't be used in these cases.
         if (includedVertices.size() <= 2) {
             return vertexPositions;
         }
 
-        final GraphMatrix l = GraphMatrix.adjacencyFromGraph(rg, includedVertices, new HashSet<>());
+        final GraphMatrix l = GraphMatrix.adjacencyFromGraph(rg, includedVertices, new IntHashSet());
 
         final EigenDecomposition e = new EigenDecomposition(MatrixUtils.createRealMatrix(l.laplacianMatrix));
         final int numVectors = e.getRealEigenvalues().length;
@@ -63,12 +62,6 @@ public class GraphSpectrumEmbedder {
 
     }
 
-    private enum MatrixType {
-
-        ADJACENCY_MATRIX,
-        LAPLACIAN_MATRIX,;
-    }
-
     private static class GraphMatrix {
 
         private final MutableIntIntMap matrixPositionToID;
@@ -81,15 +74,7 @@ public class GraphSpectrumEmbedder {
             this.dimension = laplacianMatrix.length;
         }
 
-        public static GraphMatrix adjacencyFromGraph(final GraphReadMethods rg, final Set<Integer> includedVertices, final Set<Integer> excludedLinks) {
-            return matrixFromGraph(rg, includedVertices, excludedLinks, MatrixType.ADJACENCY_MATRIX);
-        }
-
-        public static GraphMatrix laplacianFromGraph(final GraphReadMethods rg, final Set<Integer> includedVertices, final Set<Integer> excludedLinks) {
-            return matrixFromGraph(rg, includedVertices, excludedLinks, MatrixType.LAPLACIAN_MATRIX);
-        }
-
-        public static GraphMatrix matrixFromGraph(final GraphReadMethods rg, final Collection<Integer> includedVertices, final Collection<Integer> excludedLinks, final MatrixType type) {
+        public static GraphMatrix adjacencyFromGraph(final GraphReadMethods rg, final IntIterable includedVertices, final IntIterable excludedLinks) {
             final int numVertices = includedVertices.size();
             final double[][] matrixEntries = new double[numVertices][];
             for (int i = 0; i < numVertices; i++) {
@@ -111,43 +96,16 @@ public class GraphSpectrumEmbedder {
 
             for (int i = 0; i < numVertices; i++) {
                 final int vxID = matrixPositionToID.get(i);
-                int neighbourCount = 0;
                 for (int j = 0; j < rg.getVertexNeighbourCount(vxID); j++) {
                     final int neighbourID = rg.getVertexNeighbour(vxID, j);
-                    if (excludedLinks.contains(rg.getLink(vxID, neighbourID))) {
-                        continue;
-                    } else if (!includedVertices.contains(neighbourID)) {
-                        neighbourCount++;
-                        continue;
-                    } else if (idToMatrixPosition.get(neighbourID) == i) {
+                    if (excludedLinks.contains(rg.getLink(vxID, neighbourID))
+                            || !includedVertices.contains(neighbourID)
+                            || idToMatrixPosition.get(neighbourID) == i) {
                         continue;
                     }
-                    neighbourCount++;
-                    if (type != null) {
-                        switch (type) {
-                            case LAPLACIAN_MATRIX:
-                                matrixEntries[i][idToMatrixPosition.get(neighbourID)] = -1;
-                                break;
-                            case ADJACENCY_MATRIX:
-                                matrixEntries[i][idToMatrixPosition.get(neighbourID)] = 1;
-                                break;
-                            default:
-                                break;
-                        }
-                    }
+                    matrixEntries[i][idToMatrixPosition.get(neighbourID)] = 1;
                 }
-                if (type != null) {
-                    switch (type) {
-                        case LAPLACIAN_MATRIX:
-                            matrixEntries[i][i] = neighbourCount;
-                            break;
-                        case ADJACENCY_MATRIX:
-                            matrixEntries[i][i] = 0;
-                            break;
-                        default:
-                            break;
-                    }
-                }
+                matrixEntries[i][i] = 0;
             }
             return new GraphMatrix(matrixEntries, matrixPositionToID);
         }