+++ CoreArrangementPlugins/src/au/gov/asd/tac/constellation/plugins/arrangements/subgraph/SubgraphFactory.java
@@ -16,14 +16,14 @@
 package au.gov.asd.tac.constellation.plugins.arrangements.subgraph;
 
 import au.gov.asd.tac.constellation.graph.GraphWriteMethods;
-import java.util.Set;
+import org.eclipse.collections.api.set.primitive.MutableIntSet;
 
 /**
  *
  * @author twilight_sparkle
  */
+@FunctionalInterface
 public interface SubgraphFactory {
 
-    public ComponentSubgraph constructSubgraph(final GraphWriteMethods wg, final Set<Integer> includedVertexIDs);
-
+    public ComponentSubgraph constructSubgraph(final GraphWriteMethods wg, final MutableIntSet includedVertexIDs);
 }